requirements.txt
pandas
plotly
python-dotenv